app/backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy>=2.0.25,<3.0.0
alembic>=1.13.1,<2.0.0
psycopg2-binary>=2.9.9; python_version < "3.14"
psycopg[binary]>=3.2.0; python_version >= "3.14"
python-jose[cryptography]>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
bcrypt>=4.0.1,<6.0.0
python-multipart>=0.0.6,<1.0.0
pydantic>=2.5.3,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.26.0,<1.0.0
PyPDF2>=3.0.1,<4.0.0
scikit-learn>=1.4.0,<2.0.0
numpy>=1.26.3,<3.0.0
pillow>=10.2.0,<13.0.0
resend>=0.8.0,<3.0.0
pytest>=7.4.4,<9.0.0
pytest-asyncio>=0.23.3,<1.0.0
email-validator>=2.0.0,<3.0.0
